// frontend/openapi/generated/pacta/models/Analysis.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { AnalysisArtifact } from './AnalysisArtifact';
import type { AnalysisType } from './AnalysisType';
import type { FailureCode } from './FailureCode';
import type { PortfolioSnapshot } from './PortfolioSnapshot';

export type Analysis = {
    /**
     * the system assigned unique identifier of the analysis
     */
    id: string;
    /**
     * the type of analysis that was run on this object
     */
    analysisType: AnalysisType;
    /**
     * The pacta model that was used to generate this analysis
     */
    pactaVersion: string;
    /**
     * The snapshot of portfolios that was used to generate this analysis
     */
    portfolioSnapshot: PortfolioSnapshot;
    /**
     * the human meaningful name of the analysis, editable by the user
     */
    name: string;
    /**
     * Additional information about the analysis, editable by the user
     */
    description: string;
    /**
     * The time at which this analysis was created
     */
    createdAt: string;
    /**
     * The time at which this analysis was run, if set
     */
    ranAt?: string;
    /**
     * The time at which this analysis completed its run (successfully or not), if set
     */
    completedAt?: string;
    /**
     * The code describing the failure, if any
     */
    failureCode?: FailureCode;
    /**
     * The english description of the failure, if any
     */
    failureMessage?: string;
    /**
     * The list of artifacts that were generated by this analysis
     */
    artifacts: Array<AnalysisArtifact>;
    /**
     * the id of the owner of the analysis, if the user has access to see it
     */
    ownerId: string;
};